Rename withProgressBar -> index.js.
